Generated by Spark: On step 1, we need to add options that cover:
*   Artists, dancers, entertainers
*   Musicians
*   Celebrities
*   Movie stars
*   People that work in the entertainment industry

Because right now, we don't mention anything to those people. We need to make sure we create more options so people can choose from. Also, we need to create an option that says "prefer not to say" for people that might not want to disclose that information. The aware say is most popular on the business owner. We don't need to have the notification there, let's not add any notifications on step one. We should have an option that says "prefer not to say" on every page.